Fixed : breakpoints don't move when inserting or deleting lines.
Problem was that, to detect changes, the mechanism to load file contents in the textbox changed. Now the open method is used (so the IsChanged property is correctly maintained). But, doing so, the textsource of the textbox changes. So, we need to register the line_added and line_removed events everytime the file is opened.