Fixed some elements were not rebuilt when changing from debug to release builds. This ws intended in the past (only those members that had changes had to be rebuilt).
But now we should rebuild everything when switching modes, becuase the optimization setting will otherwise leave files optmized from the previous build that cannot be debugged properly. On the other hand, we only want to rebuild those files that have breakpoints, because they are the only ones that should not be optimized. Anyway, the user may want to see a variable in a non-optimized file, so it's good to rebuild everything just in case. 

This behavior could be controlled by a global setting, but for now, we play safe: builds are cleaned between changes no matter what.